Link the in/outs of post processes together to cut down copying
The post processes earlier were copying their results into the main
texture and the next post process took the task further from the main
texture. This was wasteful. Now the processes have been linked in a
way that if a post process step happens, it directly forwards its
own result buffer to the next post process to use, without copying
it into an intermittent "transfer" texture like before. Copies have been
reduced with this, with a bump of at least 8 FPS in avg in 3D scenes.
