+++ rootex/framework/systems/render_system.h
@@ -10,7 +10,6 @@
 #include "components/visual/model_component.h"
 #include "components/visual/animated_model_component.h"
 
-#include "PostProcess.h"
 #include "ASSAO/ASSAO.h"
 
 #define LINE_INITIAL_RENDER_CACHE 1000
@@ -38,31 +37,8 @@ class RenderSystem : public System
 
 	bool m_IsEditorRenderPassEnabled;
 
-	Ptr<DirectX::BasicPostProcess> m_BasicPostProcess;
-	Ptr<DirectX::DualPostProcess> m_DualPostProcess;
-	Ptr<DirectX::ToneMapPostProcess> m_ToneMapPostProcess;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_ToneMapRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_ToneMapSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_GaussianBlurRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_GaussianBlurSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_MonochromeRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_MonochromeSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_SepiaRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_SepiaSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_BloomExtractRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_BloomExtractSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_BloomHorizontalBlurRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_BloomHorizontalBlurSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_BloomVerticalBlurRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_BloomVerticalBlurSRV;
-	Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_BloomRTV;
-	Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_BloomSRV;
-
-	ASSAO_Effect* m_ASSAO;
-
 	RenderSystem();
 	RenderSystem(RenderSystem&) = delete;
-	virtual ~RenderSystem();
 
 	void renderPassRender(float deltaMilliseconds, RenderPass renderPass);
